+++ buildConfig.js
@@ -4,8 +4,24 @@ module.exports = {
     path: path.join(__dirname, 'ui-test'),
     testEntryPoint: path.join(__dirname, 'test', 'index.jsx'),
     webpackConfig: {
-        debug: require('./webpack.config.js'),
-        production: require('./webpack.config.prod.js'),
+        common: {
+            context: path.resolve(__dirname),
+        },
+        debug: {
+            entry: './ui-test/ui-test.jsx',
+            output: {
+                path: path.join(__dirname, 'ui-test'),
+                filename: 'component.min.js',
+            },
+        },
+        production: {
+            entry: './index.js',
+            output: {
+                path: path.join(__dirname, 'es5'),
+                filename: 'component.js',
+                libraryTarget: 'commonjs2',
+            },
+        },
     },
     lintingFiles: [
         './src/**/*',